+++ CompanionPackages~/com.playserv.google-signin/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.3.4] - 2026-07-25
+
+- Aligned package and core SDK dependency versions with `0.3.4`.
+
 ## [0.3.3] - 2026-07-25
 
 - Aligned package and core SDK dependency versions with `0.3.3`.